Performance improvement
Speed up initial rendering by loading submission-wizard-goal asynchronously.
