+++ app/elements/submission-wizard-goal.ts
@@ -4,7 +4,7 @@ import { localized, msg } from '@lit/localize';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 
 import resetScroll from '../directives/reset-scroll';
-import { Goal, Section, findGoal } from '../data';
+import { Goal, Section, findGoal } from '../data/goal';
 import { LocalizationMixin } from '../localization';
 
 import baseStyle from './base.css';
